+++ worldguard-bukkit/src/main/java/com/sk89q/worldguard/bukkit/listener/EventAbstractionListener.java
@@ -163,6 +163,7 @@ import javax.annotation.Nullable;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
 public class EventAbstractionListener extends AbstractListener {
@@ -209,8 +210,21 @@ public class EventAbstractionListener extends AbstractListener {
         }
     }
 
+    private boolean isExemptBlock(Material material) {
+        // Generating an End Portal from Bedrock/End_Portal_Frame should not trigger BlockMultiPlaceEvent
+        // Canceling this event for these blocks causes an upstream item duplication bug.
+        // https://github.com/PaperMC/Paper/issues/13586
+        return switch (material) {
+            case BEDROCK, END_PORTAL_FRAME -> true;
+            default -> false;
+        };
+    }
+
     @EventHandler(ignoreCancelled = true)
     public void onBlockMultiPlace(BlockMultiPlaceEvent event) {
+        if (isExemptBlock(event.getBlockPlaced().getType())) {
+            return;
+        }
         List<Block> placed = event.getReplacedBlockStates().stream().map(BlockState::getBlock).collect(Collectors.toList());
         int origAmt = placed.size();
         PlaceBlockEvent delegateEvent = new PlaceBlockEvent(event, create(event.getPlayer()), event.getBlock().getWorld(),
@@ -1008,6 +1022,7 @@ public class EventAbstractionListener extends AbstractListener {
     }
 
     @EventHandler(ignoreCancelled = true)
+    @SuppressWarnings({"rawtypes", "unchecked"})
     public void onInventoryMoveItem(InventoryMoveItemEvent event) {
         InventoryHolder causeHolder = PaperLib.getHolder(event.getInitiator(), false).getHolder();
 
@@ -1044,9 +1059,19 @@ public class EventAbstractionListener extends AbstractListener {
                 handleInventoryHolderUse(event, cause, targetHolder);
             }
 
-            if (event.isCancelled() && causeHolder instanceof Hopper && wcfg.breakDeniedHoppers) {
-                Bukkit.getScheduler().scheduleSyncDelayedTask(getPlugin(),
-                        () -> ((Hopper) causeHolder).getBlock().breakNaturally());
+            if (event.isCancelled() && causeHolder instanceof Hopper hopper && wcfg.breakDeniedHoppers) {
+                Runnable task = () -> hopper.getBlock().breakNaturally();
+
+                if (WorldGuardPlugin.inst().isFolia()) {
+                    Bukkit.getRegionScheduler().run(getPlugin(), hopper.getLocation(), new Consumer() {
+                        @Override
+                        public void accept(Object ignored) {
+                            task.run();
+                        }
+                    });
+                } else {
+                    Bukkit.getScheduler().scheduleSyncDelayedTask(getPlugin(), task);
+                }
             } else {
                 entry.setCancelled(event.isCancelled());
             }