+++ worldguard-bukkit/src/main/java/com/sk89q/worldguard/bukkit/util/report/SchedulerReport.java
@@ -24,6 +24,7 @@ import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
 import com.sk89q.worldedit.util.report.DataReport;
+import com.sk89q.worldguard.bukkit.WorldGuardPlugin;
 import org.bukkit.Bukkit;
 import org.bukkit.scheduler.BukkitTask;
 
@@ -51,6 +52,9 @@ public class SchedulerReport extends DataReport {
 
     public SchedulerReport() {
         super("Scheduler");
+        if (WorldGuardPlugin.inst().isFolia()) {
+            return; // Todo Add support Schedulers
+        }
 
         List<BukkitTask> tasks = Bukkit.getServer().getScheduler().getPendingTasks();